separates api logic from main
