+++ api/__init__.py
@@ -1,0 +1,5 @@
+from sanic import Blueprint
+
+from .cases import cases
+
+api = Blueprint.group(cases, url_prefix="/api")
